add context manager for file reading and remove useless function from the fileio module
